Complete LC Analysis
General version of my previous code LCAnalysis_2. This new version does
the following:

1) It uploads a file containing photometric data and plots the raw LC.

2) With an approximate measure of t0 taken from the archive, it marks
the position of the transits with red vertical lines.

3) Using the results of the code LCAnalysis_1 (available on Github), it
studies every transit — one by one—, and then finds the best estimate
for the time of transit.

4) It plots a theoretical LC for every transit in the data.

5) It utilises the EMCEE library to compute the uncertainties of the
new transit times.

6) Output: Multiple figures and a text file containing the final
results (for each transit, the best guess for the transit time along
with its upper and lower errors).