Complete LC Analysis with 2 free parameters (t0, f0)
Additions to LCAnalysis_3:

1. We now let the out-of-transit flux (f0) and the time of transit (t0)
be our free parameters. For each variable, we find the best fit and its
upper and lower error.

2. We do a linear fit to the observed times of transit to recalculate
the period and t0. Note: We don’t take into consideration the upper and
lower errors of tObs. In addition, the fit does not yield the errors of
the new period and t0. This is something to be corrected in the future.

3. We plot t_obs[n] vs. t_obs[n]-t0-P*n and include the uncertainties
of t_obs derived from the MCMC analysis.